Move reducer to another file
